Update mapping-hisat2-single-end.ga
I made some changes to the descriptions for each tool to fix a few grammatical mistakes, unify the descriptions to look consistent and to make the transition between steps more meaningful to a user who does not know this is a Galaxy Workflow.  I don't think I flubbed up anything with it but  I recommend testing it just in case.